refactor: move sandbox-related components out of editor folder
Move terminal, vnc-browser, web-preview, mobile-preview, secrets, and
shared components from components/editor to components/sandbox.

The editor folder now only contains code editing related components
(code-view, editor-core, editor-view, file-preview, file-tree), while
sandbox runtime/preview features live in the new sandbox folder.
