+++ frontend/src/components/sandbox/mobile-preview/MobilePreview.tsx
