+++ frontend/src/components/sandbox/secrets/SecretsView.tsx
