+++ frontend/src/components/sandbox/shared/NoOpenPortsState.tsx
