+++ frontend/src/components/sandbox/terminal/TerminalPanel.tsx
