+++ frontend/src/components/sandbox/terminal/TerminalTab.tsx
