+++ frontend/src/components/sandbox/vnc-browser/VNCClient.tsx
