+++ frontend/src/components/sandbox/web-preview/Panel.tsx
